commit to allow for pull

+++ loans.h
@@ -47,7 +47,7 @@
  #ifndef CALCULATE_INTEREST_DEF
   float Calculate_Interest(struct loan *account,
 			   uint8_t account_number,
-			   uint16_t number_of_days);
+			   uint8_t number_of_days);
   #define CALCULATE_INTEREST_DEF
  #endif
